make base.py thinner, add theme.py, session.py and routes.py
